refactor: moved status circles to its own folder component
