+++ src/components/Web3Modals/components/Transaction/Transaction.tsx
@@ -1,6 +1,6 @@
 import { FiArrowUpRight, FiCheckCircle, FiXCircle } from 'react-icons/fi';
 import { Transaction as TransactionInterface } from 'types/types.guilds';
-import { PendingCircle } from 'components/primitives/PendingCircle';
+import { PendingCircle } from 'components/primitives/StatusCircles';
 import { TransactionContainer, Link, Icon } from './Transaction.styled';
 import { getBlockExplorerUrl } from 'provider';
 import { useNetwork } from 'wagmi';
